# waterwind-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WaterWind API",
        "description": "WaterWind – Smart Water Recommendation Platform REST API. Authenticate via Sanctum Bearer tokens.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/crops": {
            "get": {
                "tags": [
                    "Crops"
                ],
                "summary": "List all crops with optional search",
                "operationId": "6fc8af3094fd521638fd4d265d8bd106",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filter by name, category, or device",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Crops retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Crops"
                ],
                "summary": "Create a new crop",
                "operationId": "6bf944c4823263e051225784ea3903d6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "recommended_water",
                                    "recommended_ph",
                                    "recommended_device"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tomato"
                                    },
                                    "scientific_name": {
                                        "type": "string",
                                        "example": "Solanum lycopersicum"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Vegetable"
                                    },
                                    "recommended_ph": {
                                        "type": "number",
                                        "example": 6.2
                                    },
                                    "recommended_ph_min": {
                                        "type": "number",
                                        "example": 6
                                    },
                                    "recommended_ph_max": {
                                        "type": "number",
                                        "example": 6.8
                                    },
                                    "recommended_water": {
                                        "type": "number",
                                        "example": 400
                                    },
                                    "recommended_water_type": {
                                        "type": "string",
                                        "example": "Fresh Water"
                                    },
                                    "irrigation_frequency": {
                                        "type": "string",
                                        "example": "Daily"
                                    },
                                    "water_requirement_liters_per_hectare": {
                                        "type": "number",
                                        "example": 40000
                                    },
                                    "growth_period_days": {
                                        "type": "integer",
                                        "example": 80
                                    },
                                    "temperature_min": {
                                        "type": "number",
                                        "example": 18
                                    },
                                    "temperature_max": {
                                        "type": "number",
                                        "example": 29
                                    },
                                    "humidity_min": {
                                        "type": "number",
                                        "example": 65
                                    },
                                    "humidity_max": {
                                        "type": "number",
                                        "example": 85
                                    },
                                    "wind_min": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "wind_max": {
                                        "type": "number",
                                        "example": 4
                                    },
                                    "recommended_device": {
                                        "type": "string",
                                        "example": "WaterWind Micro"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Crop created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "409": {
                        "description": "Duplicate crop name"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/crops/{id}": {
            "put": {
                "tags": [
                    "Crops"
                ],
                "summary": "Update a crop",
                "operationId": "d9f5183aeaf220fe05d359735fd79a93",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Crop updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Crop not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Crops"
                ],
                "summary": "Soft-delete a crop",
                "operationId": "04e844b0c8bd801c8ef582aee533ad20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Crop deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Crop not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/crops/{id}/soils": {
            "post": {
                "tags": [
                    "Crops"
                ],
                "summary": "Sync compatible soils for a crop",
                "operationId": "5595521998fcf6869de77bfad6aa2bcb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "soil_ids"
                                ],
                                "properties": {
                                    "soil_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            3,
                                            5
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compatible soils updated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Crop not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/crops/{id}/minerals": {
            "post": {
                "tags": [
                    "Crops"
                ],
                "summary": "Sync minerals for a crop with priority levels",
                "operationId": "f5465df415a051e7f0404dcd77ef9017",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "minerals"
                                ],
                                "properties": {
                                    "minerals": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "mineral_id": {
                                                    "type": "integer"
                                                },
                                                "priority": {
                                                    "type": "string",
                                                    "enum": [
                                                        "low",
                                                        "medium",
                                                        "high"
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Minerals updated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Crop not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token as: Bearer {token}",
                "scheme": "bearer"
            }
        }
    },
...
}